pyproject.toml
[tool.poetry]
name = "architecturecourse"
version = "0.1.0"
description = ""
authors = ["jblanquicett92 <jorgeabm1992@gmail.com>"]

[tool.poetry.dependencies]
wheel = "0.37.1"
python = "^3.8"
sqlalchemy = "1.4.9"
flask = "2.0"
Flask-SQLAlchemy = "2.5.1"
psycopg2-binary = "2.9.3"
redis = "4.1.1"
python-dotenv = "0.19.2"
marshmallow = "3.14.1"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
pytest-icdiff = "0.5"
mypy = "0.931"
pylint ="2.12.2"
requests = "2.26.0"
black = {version = "^22.1.0", allow-prereleases = true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
